pyproject.toml
[project]
name = "llm-inference"
version = "0.1.0"
description = "Add your description here"
readme = "Readme.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.0.15",
    "diskcache>=5.6.3",
    "fastapi>=0.116.1",
    "huggingface-hub>=0.33.4",
    "langchain-community>=0.3.27",
    "langchain-huggingface>=0.3.0",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.3",
    "pydantic-settings>=2.10.1",
    "pytest-cov>=6.2.1",
    "selectolax>=0.3.31",
    "sentence-transformers>=5.0.0",
    "sqlalchemy>=2.0.41",
    "sse-starlette>=2.4.1",
    "starlette-context>=0.4.0",
    "unstructured>=0.18.7",
    "uvicorn[standard]>=0.35.0",
]

[tool.black]
line-length = 88
target-version = ['py311']
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.12.3",
]
